// src/wmmt/v388.proto.d.ts
export namespace v388 {

    /** Namespace protobuf. */
    namespace protobuf {
// ...
        namespace SaveGameResultRequest {
// ...
            interface ITimeAttackResult {

                /** TimeAttackResult time */
                time: number;

                /** TimeAttackResult course */
                course: number;

                /** TimeAttackResult section_1Time */
                section_1Time: number;

                /** TimeAttackResult section_2Time */
                section_2Time: number;

                /** TimeAttackResult section_3Time */
                section_3Time: number;

                /** TimeAttackResult section_4Time */
                section_4Time: number;

                /** TimeAttackResult section_5Time */
                section_5Time?: (number|null);

                /** TimeAttackResult section_6Time */
                section_6Time?: (number|null);

                /** TimeAttackResult section_7Time */
                section_7Time?: (number|null);
            }
// ...
        }
// ...
    }
}